% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonrewind-package.R
\docType{package}
\name{geojsonrewind-package}
\alias{geojsonrewind-package}
\alias{geojsonrewind}
\title{\strong{GeoJSON Rewinding}}
\description{
a package for fixing ring order, see also the
right-hand rule for geojson
}
\author{
Scott Chamberlain \email{sckott@protonmail.com}
}
