% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewind.R
\name{rewind}
\alias{rewind}
\title{rewind}
\usage{
rewind(x, outer = TRUE)
}
\arguments{
\item{x}{A geojson object, either as list, character string, or json
class}

\item{outer}{(logical) clockwise (\code{TRUE}, default) or
counterclockwise (\code{FALSE})}
}
\value{
a geojson object, as json/character class
}
\description{
rewind
}
\details{
ported from \code{geojson-rewind} JS library at
\url{https://github.com/mapbox/geojson-rewind}
}
\examples{
x <- '{
 "type":"Polygon",
 "coordinates":[
   [[100.0,0.0],[100.0,1.0],[101.0,1.0],[101.0,0.0],[100.0,0.0]]
 ]
}'
rewind(x)
rewind(x, outer = FALSE)
}

