% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ring_area.R
\name{ring_area}
\alias{ring_area}
\title{geojson area function}
\usage{
ring_area(x)
}
\arguments{
\item{x}{A list of coordinates. See examples for its form}
}
\value{
single integer - The approximate signed geodesic area of the
polygon in square meters. Positive if ring is oriented
clockwise, otherwise it will be negative.
}
\description{
geojson area function
}
\details{
ported from \code{geojson-aera} JS library at
\url{https://github.com/mapbox/geojson-area}

Calculate the approximate area of the polygon were it projected onto
the earth.  Note that this area will be positive if ring is oriented
clockwise, otherwise it will be negative.

Reference:
Robert. G. Chamberlain and William H. Duquette, "Some Algorithms for
Polygons on a Sphere", JPL Publication 07-03, Jet Propulsion
Laboratory, Pasadena, CA, June 2007
http://trs-new.jpl.nasa.gov/dspace/handle/2014/40409
}
\examples{
x <- '{
 "type":"Polygon",
 "coordinates":[
   [[100.0,0.0],[100.0,1.0],[101.0,1.0],[101.0,0.0],[100.0,0.0]]
 ]
}'
x <- jsonlite::fromJSON(x, FALSE)
ring_area(x = x$coordinates[[1]])
ring_area(x = rev(x$coordinates[[1]]))
}
